perf: optimizing parameters sizing (#207)
* perf: optimizing parameters sizing

* chore: feedback


* style: formatting

* perf: more optimizations

* chore: fixes

* chore: fix certora spec

* chore: more fixes for certora spec

* chore: more and more fixes for certora spec

* fix: ends type

* test(certora): timestamp conversion

* test(certora): timestamp conversion again

* test(certora): timestamp conversion revert to assert_uint64

* test(certora): timestamp with mathint

* test(certora): timestamp back with uint64 with require

* Add missing configuration

* Fix previous merge

* Update StorageRequested to use int64 for expiry

* requestDurationLimit => uint64

---------

+++ certora/harness/MarketplaceHarness.sol
@@ -14,7 +14,7 @@ contract MarketplaceHarness is Marketplace {
         Marketplace(config, token, verifier)
     {}
 
-    function publicPeriodEnd(Period period) public view returns (uint256) {
+    function publicPeriodEnd(Period period) public view returns (uint64) {
         return _periodEnd(period);
     }
 
@@ -22,7 +22,7 @@ contract MarketplaceHarness is Marketplace {
         return _slots[slotId];
     }
 
-    function generateSlotId(RequestId requestId, uint256 slotIndex) public pure returns (SlotId) {
+    function generateSlotId(RequestId requestId, uint64 slotIndex) public pure returns (SlotId) {
         return Requests.slotId(requestId, slotIndex);
     }
 }
